break up verifierAlgorithm into separate files bases on Curve
